backend/requirements-prod.txt
asyncpg==0.30.0 
cryptography==44.0.2
Fernet==1.0.1
Flask==3.1.0
flask-cors==5.0.1
flask[async]
loguru==0.7.3
PyMuPDF==1.24.14
PyPDF2==3.0.1
uwsgi
python-dotenv===1.0.1
reportlab==4.3.1
swagger-ui-py==23.9.23
tabulate==0.9.0
werkzeug==3.1.3

